// src/channels/wechat/api/types.ts
export type BaseInfo = {
  channel_version?: string
}

export const MessageType = {
  NONE: 0,
  USER: 1,
  BOT: 2,
} as const

export const MessageItemType = {
  NONE: 0,
  TEXT: 1,
  IMAGE: 2,
  VOICE: 3,
  FILE: 4,
  VIDEO: 5,
} as const

export const MessageState = {
  NEW: 0,
  GENERATING: 1,
  FINISH: 2,
} as const

export type TextItem = {
  text?: string
}

export type CDNMedia = {
  encrypt_query_param?: string
  aes_key?: string
  encrypt_type?: number
  full_url?: string
}

export type ImageItem = {
  media?: CDNMedia
  thumb_media?: CDNMedia
  aeskey?: string
  url?: string
  mid_size?: number
  thumb_size?: number
  thumb_height?: number
  thumb_width?: number
  hd_size?: number
}

export type VoiceItem = {
  media?: CDNMedia
  encode_type?: number
  bits_per_sample?: number
  sample_rate?: number
  playtime?: number
  text?: string
}

export type FileItem = {
  media?: CDNMedia
  file_name?: string
  md5?: string
  len?: string
}

export type VideoItem = {
  media?: CDNMedia
  video_size?: number
  play_length?: number
  video_md5?: string
  thumb_media?: CDNMedia
  thumb_size?: number
  thumb_height?: number
  thumb_width?: number
}

export type RefMessage = {
  message_item?: MessageItem
  title?: string
}

export type MessageItem = {
  type?: number
  create_time_ms?: number
  update_time_ms?: number
  is_completed?: boolean
  msg_id?: string
  ref_msg?: RefMessage
  text_item?: TextItem
  image_item?: ImageItem
  voice_item?: VoiceItem
  file_item?: FileItem
  video_item?: VideoItem
}

export type WechatMessage = {
  seq?: number
  message_id?: number
  from_user_id?: string
  to_user_id?: string
  client_id?: string
  create_time_ms?: number
  update_time_ms?: number
  delete_time_ms?: number
  session_id?: string
  group_id?: string
  message_type?: number
  message_state?: number
  item_list?: MessageItem[]
  context_token?: string
}

export type GetUpdatesReq = {
  sync_buf?: string
  get_updates_buf?: string
}

export type GetUpdatesResp = {
  ret?: number
  errcode?: number
  errmsg?: string
  msgs?: WechatMessage[]
  sync_buf?: string
  get_updates_buf?: string
  longpolling_timeout_ms?: number
}

export type SendMessageReq = {
  msg?: WechatMessage
}

export type SendMessageResp = Record<string, never>

export const TypingStatus = {
  TYPING: 1,
  CANCEL: 2,
} as const

export type SendTypingReq = {
  ilink_user_id?: string
  typing_ticket?: string
  status?: number
}

export type SendTypingResp = {
  ret?: number
  errmsg?: string
}

export type GetConfigResp = {
  ret?: number
  errmsg?: string
  typing_ticket?: string
}

export type NotifyStopResp = {
  ret?: number
  errmsg?: string
}

export type NotifyStartResp = {
  ret?: number
  errmsg?: string
}